requirements.txt
streamlit>=1.30.0
python-docx>=1.1.0
gspread>=6.0.0
google-auth>=2.25.0
